chore: added support to ServerInfo API and refactored the planner model
