+++ src/edu/usf/cutr/otp/planner/model/From.kt
@@ -1,11 +1,11 @@
-package edu.usf.cutr.otp.model
+package edu.usf.cutr.otp.planner.model
 
 
-data class To_(
+data class From
+(
     val name: String? = null,
     val lon: Double? = null,
     val lat: Double? = null,
-    val arrival: Int? = null,
     val vertexType: String? = null,
     val additionalProperties: MutableMap<String, Any> = HashMap()
 )
