+++ src/edu/usf/cutr/otp/planner/model/From_.kt
@@ -1,11 +1,12 @@
-package edu.usf.cutr.otp.model
+package edu.usf.cutr.otp.planner.model
 
 
-data class From
+data class From_
 (
     val name: String? = null,
     val lon: Double? = null,
     val lat: Double? = null,
+    val departure: Int? = null,
     val vertexType: String? = null,
     val additionalProperties: MutableMap<String, Any> = HashMap()
 )
