+++ src/edu/usf/cutr/otp/planner/model/Plan.kt
@@ -1,6 +1,6 @@
-package edu.usf.cutr.otp.model
+package edu.usf.cutr.otp.planner.model
 
-import java.util.HashMap
+import java.util.*
 
 
 data class Plan
