+++ src/edu/usf/cutr/otp/planner/model/To_.kt
@@ -1,12 +1,11 @@
-package edu.usf.cutr.otp.model
+package edu.usf.cutr.otp.planner.model
 
 
-data class From_
-(
+data class To_(
     val name: String? = null,
     val lon: Double? = null,
     val lat: Double? = null,
-    val departure: Int? = null,
+    val arrival: Int? = null,
     val vertexType: String? = null,
     val additionalProperties: MutableMap<String, Any> = HashMap()
 )
