+++ src/edu/usf/cutr/otp/serverinfo/model/ServerInfo.kt
@@ -1,0 +1,10 @@
+package edu.usf.cutr.otp.serverinfo.model
+
+import java.util.*
+
+data class ServerInfo (
+    var serverVersion: ServerVersion? = null,
+    var cpuName: String? = null,
+    var nCores: Int? = null ,
+    val additionalProperties: MutableMap<String, Any> = HashMap()
+)
